pyproject.toml
[project]
name = "micromdcodehighlighter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0.2",
    "ruamel-yaml>=0.18.12",
]
